+++ src/TitleDetails/ExpectedPiecesList/ExpectedPiecesList.js
@@ -1,37 +1,43 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 import {
   PIECE_STATUS,
 } from '@folio/stripes-acq-components';
 
-import { usePaginatedPieces } from '../../common/hooks';
-
 import {
   PIECE_COLUMNS,
 } from '../constants';
+import {
+  usePiecesList,
+} from '../hooks';
 import PiecesList from '../PiecesList';
 
-const RESULT_COUNT_INCREMENT = 30;
-
-const ExpectedPiecesList = ({ title, selectPiece, visibleColumns }) => {
-  const [sorting, setSorting] = useState({
+const ExpectedPiecesList = ({
+  filters,
+  onLoadingStatusChange,
+  title,
+  selectPiece,
+  visibleColumns,
+}) => {
+  const initialSorting = {
     sorting: 'receiptDate',
     sortingDirection: 'ascending',
-  });
-  const [pagination, setPagination] = useState({ limit: RESULT_COUNT_INCREMENT, offset: 0, timestamp: new Date() });
+  };
+
   const {
-    pieces,
-    totalRecords,
     isFetching,
-  } = usePaginatedPieces({
     pagination,
-    queryParams: {
-      titleId: title.id,
-      poLineId: title.poLineId,
-      receivingStatus: PIECE_STATUS.expected,
-      ...sorting,
-    },
+    pieces,
+    setPagination,
+    setSorting,
+    totalRecords,
+  } = usePiecesList({
+    filters,
+    initialSorting,
+    onLoadingStatusChange,
+    title,
+    queryParams: { receivingStatus: PIECE_STATUS.expected },
   });
 
   return (
@@ -50,6 +56,8 @@ const ExpectedPiecesList = ({ title, selectPiece, visibleColumns }) => {
 };
 
 ExpectedPiecesList.propTypes = {
+  filters: PropTypes.object.isRequired,
+  onLoadingStatusChange: PropTypes.func.isRequired,
   title: PropTypes.object.isRequired,
   selectPiece: PropTypes.func.isRequired,
   visibleColumns: PropTypes.arrayOf(PropTypes.string).isRequired,