+++ src/TitleDetails/ReceivedPiecesList/ReceivedPiecesList.js
@@ -1,35 +1,39 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 import {
   PIECE_STATUS,
 } from '@folio/stripes-acq-components';
 
-import { usePaginatedPieces } from '../../common/hooks';
-
 import { PIECE_COLUMNS } from '../constants';
 import PiecesList from '../PiecesList';
+import { usePiecesList } from '../hooks';
 
-const RESULT_COUNT_INCREMENT = 30;
-
-const ReceivedPiecesList = ({ title, selectPiece, visibleColumns }) => {
-  const [sorting, setSorting] = useState({
+const ReceivedPiecesList = ({
+  filters,
+  onLoadingStatusChange,
+  title,
+  selectPiece,
+  visibleColumns,
+}) => {
+  const initialSorting = {
     sorting: 'receivedDate',
     sortingDirection: 'descending',
-  });
-  const [pagination, setPagination] = useState({ limit: RESULT_COUNT_INCREMENT, offset: 0, timestamp: new Date() });
+  };
+
   const {
-    pieces,
-    totalRecords,
     isFetching,
-  } = usePaginatedPieces({
     pagination,
-    queryParams: {
-      titleId: title.id,
-      poLineId: title.poLineId,
-      receivingStatus: PIECE_STATUS.received,
-      ...sorting,
-    },
+    pieces,
+    setPagination,
+    setSorting,
+    totalRecords,
+  } = usePiecesList({
+    filters,
+    initialSorting,
+    onLoadingStatusChange,
+    title,
+    queryParams: { receivingStatus: PIECE_STATUS.received },
   });
 
   return (
@@ -48,6 +52,8 @@ const ReceivedPiecesList = ({ title, selectPiece, visibleColumns }) => {
 };
 
 ReceivedPiecesList.propTypes = {
+  filters: PropTypes.object.isRequired,
+  onLoadingStatusChange: PropTypes.func.isRequired,
   title: PropTypes.object.isRequired,
   selectPiece: PropTypes.func.isRequired,
   visibleColumns: PropTypes.arrayOf(PropTypes.string).isRequired,