+++ src/TitleDetails/hooks/usePiecesList/index.js
@@ -1,0 +1,1 @@
+export { usePiecesList } from './usePiecesList';